move : `art.py` splitted into `functions.py` and `utils.py`.
