+++ art/__init__.py
@@ -1,10 +1,11 @@
 # -*- coding: utf-8 -*-
 """Art modules."""
-from .art import artError
-from .art import aprint, art, randart
-from .art import tprint, tsave, text2art
-from .art import decor
-from .art import get_font_dic, set_default, help_func, art_list, font_list, decor_list
+from .errors import artError
+from .functions import aprint, art, randart
+from .functions import tprint, tsave, text2art
+from .functions import decor
+from .functions import get_font_dic, set_default, art_list, font_list, decor_list
+from .__main__ import help_func
 from .params import ART_VERSION, FONT_NAMES, ART_NAMES, DECORATION_NAMES, DEFAULT_FONT
 from .params import ART_COUNTER, FONT_COUNTER, DECORATION_COUNTER
 from .params import NON_ASCII_ARTS, NON_ASCII_FONTS, ASCII_FONTS, ASCII_ARTS
